refactor(blobstore-s3): clean-up configuration
